blizzard: clean up faucet wallet finds
